Configuration management

+++ 0x0A-configuration_management/1-install_a_package.pp
@@ -6,5 +6,7 @@ package { 'python3-pip':
 # Install Flask version 2.1.0
 exec { 'install_flask':
   command => '/usr/bin/pip3 install flask==2.1.0', # Command to install Flask
+  path    => ['/usr/bin', '/usr/sbin'],
+  unless  => '/usr/bin/pip3 show flask | grep -q "Version: 2.1.0"', # Prevent reinstallation if version 2.1.0 is already installed
   require => Package['python3-pip'], # Ensure this runs after pip3 is installed
 }
